--- app/Http/Controllers/Controllers/Api/HandleImageUploadController.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Api;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Storage;
-
-class HandleImageUploadController extends Controller
-{
-    /**
-     * Handle the incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function __invoke(Request $request)
-    {
-        $data = $request->allFiles();
-
-        foreach ($data as $img) {
-            Storage::put('uploads', $img);
-        }
-
-        return response()->json('All files uploaded', 200);
-    }
-}
