--- app/Http/Controllers/Controllers/Auth/ResetPasswordController.php
@@ -1,87 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Foundation\Auth\ResetsPasswords;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Http\Request;
-use App\User;
-use Auth;
-use DB;
-
-class ResetPasswordController extends Controller
-{
-    /*
-    |--------------------------------------------------------------------------
-    | Password Reset Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller is responsible for handling password reset requests
-    | and uses a simple trait to include this behavior. You're free to
-    | explore this trait and override any methods you wish to tweak.
-    |
-    */
-
-    use ResetsPasswords;
-
-    /**
-     * Where to redirect users after resetting their password.
-     *
-     * @var string
-     */
-    protected $redirectTo = '/home';
-
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware('guest', ['except' => 'changePassword']);
-    }
-
-    /**
-     * Get a validator for an incoming registration request.
-     *
-     * @param  array  $data
-     * @return \Illuminate\Contracts\Validation\Validator
-     */
-    protected function validator(array $data)
-    {
-        return Validator::make($data, [
-            'password' => 'required|min:6',
-            'password_confirmation' => 'same:password'
-        ]);
-    }
-
-    /**
-     * Create a new user instance after a valid registration.
-     *
-     * @param  array  $data
-     * @return User
-     */
-    protected function update(array $data)
-    {
-        $user = User::find(Auth::user()->id);
-        $user->password = bcrypt($data['password']);
-
-        return $user->save();
-    }
-
-    /**
-     * Save new user password.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     */
-    public function changePassword(Request $request)
-    {
-        if ($this->validator($request->all())->fails()) {
-            return $this->validator($request->all())->errors();
-        }
-
-        $user = $this->update($request->all());
-        return "You've successfully change your password.";
-    }
-}
