--- app/Http/Controllers/Controllers/DownloadFileController.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
-
-class DownloadFileController extends Controller
-{
-    /**
-     * Handle the incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Symfony\Component\HttpFoundation\StreamedResponse
-     */
-    public function __invoke(Request $request, $filePath)
-    {
-        return Storage::download($filePath);
-    }
-}
