+++ app/Http/Controllers/InputController.php
@@ -2,14 +2,20 @@
 
 namespace App\Http\Controllers;
 
+use App\Imports\CommodityDataImport;
+use App\Imports\PopulationDataImport;
 use App\Models\Crop;
 use App\Models\Population;
 use App\Models\SrcCommodity;
 use App\Models\SrcProvince;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Http;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use Maatwebsite\Excel\Facades\Excel;
 use PhpOffice\PhpSpreadsheet\Reader\Csv;
 
 class InputController extends Controller
@@ -101,189 +107,104 @@ class InputController extends Controller
         // identify type of crop
         $typeCrop = SrcCommodity::find($commodityId);
 
-        // check for errors in commodity data
-        $error1 = $this->checkCommodityData($commodityDest, $typeCrop->crop_type);
+//        dd($request->except('crop_data_filename','pop_data_filename'));
 
-        // check for errors in pop growth rate
-        $error2 = $this->checkPopGrowthRate($popGrowthDest, $year);
-
-        // return errors if any
-        if ($error1) {
-            return $error1;
-        } elseif ($error2) {
-            return $error2;
-        }
-
-        // convert percent to decimal value
-        $conversionRate = $conversionRate / 100;
-
-        // save crop details
-        $crop = Crop::create($request->except('crop_data_filename','pop_data_filename'));
-        $crop->crop_data_filename = $commodityDest;
-        $crop->pop_data_filename = $popGrowthDest;
-        $crop->save();
-
-        // save population
-        $population = Population::create([
-            'crop_id' => $crop->id,
+        // create crop
+        $crop = auth()->user()->crops()->create([
+            'src_commodity_id' => $request->commodity_id,
+            'conversion_rate' => $request->conversion_rate,
+            'src_province_id' => $request->province_id,
+            'population' => $request->population,
             'year' => $request->year,
-            'population' => $request->population
+            'per_capita' => $request->per_capita,
+            'per_capita_year' => $request->per_capita_year,
+            'commodity_data' => $request->commodity_data,
+            'pop_growth_rate' => $request->pop_growth_rate,
+            'crop_data_filename' => $commodityDest,
+            'pop_data_filename' => $popGrowthDest
         ]);
+
+        $this->importCommodityData($commodityData, $crop->id, $conversionRate, $typeCrop->crop_type);
+
+        // TODO: Save data to database
+        $this->handlePopulationProjection($population, $crop->id, $request->file('pop_growth_rate'));
+
+        // run models in python
+        $this->runModelsInPython($commodityData, $crop->id, $conversionRate, $typeCrop->crop_type);
 
         // import commodity data
         $this->importCommodityData($commodityData, $crop->id, $conversionRate, $typeCrop->crop_type);
 
-        if ($typeCrop->crop_type == 'crops') {
-
-            // compute slope
-            $slopeArea = $this->calculateSlope($crop->id, 'crop_data', 'ln_area');
-            $slopeYield = $this->calculateSlope($crop->id, 'crop_data', 'ln_yield');
-            $slopeConsumption = $this->calculateSlope($crop->id, 'crop_data', 'ln_consumption');
-
-            // save slope
-            DB::table('crop_slope')
-                ->insert([
-                    'crop_id' => $crop->id,
-                    'area' => $slopeArea,
-                    'yield' => $slopeYield,
-                    'consumption' => $slopeConsumption
-                ]);
-
-            // compute annualize growth rate
-            $agrArea = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'harvested_area_ha');
-            $agrYield = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'yield_mt_ha');
-            $agrConsumption = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'per_capita_consumption_kg_yr');
-
-            // save anualize growth rate
-            DB::table('crop_annualized_growth_rate')
-                ->insert([
-                    'crop_id' => $crop->id,
-                    'area' => $agrArea,
-                    'yield' => $agrYield,
-                    'consumption' => $agrConsumption
-                ]);
-
-        } elseif ($typeCrop->crop_type == 'Non-crops') {
-
-            // compute slope
-            $slopeProduction = $this->calculateSlope($crop->id, 'non_crop_data', 'ln_production');
-            $slopeConsumption = $this->calculateSlope($crop->id, 'non_crop_data', 'ln_consumption');
-
-            // save slope
-            DB::table('non_crop_slope')
-                ->insert([
-                    'crop_id' => $crop->id,
-                    'production' => $slopeProduction,
-                    'consumption' => $slopeConsumption
-                ]);
-
-            // compute annualize growth rate
-            $agrProduction = $this->calculateAnnGrowthRate($crop->id, 'non_crop_data', 'production_mt');
-            $agrConsumption = $this->calculateAnnGrowthRate($crop->id, 'non_crop_data', 'per_capita_consumption_kg_yr');
-
-            // save anualize growth rate
-            DB::table('non_crop_annualized_growth_rate')
-                ->insert([
-                    'crop_id' => $crop->id,
-                    'production' => $agrProduction,
-                    'consumption' => $agrConsumption
-                ]);
-
-        }
-
-        // import population growth rate
-        $this->importPopGrowthRate($popGrowthDest, $crop->id, $population);
-
-        // create file source for auto-arima model
-        $this->createAutoArimaSrcData($crop->id, $typeCrop->crop_type);
+        // the codes below will be run in python
+//        if ($typeCrop->crop_type == 'crops') {
+//
+//            // compute slope
+//            $slopeArea = $this->calculateSlope($crop->id, 'crop_data', 'ln_area');
+//            $slopeYield = $this->calculateSlope($crop->id, 'crop_data', 'ln_yield');
+//            $slopeConsumption = $this->calculateSlope($crop->id, 'crop_data', 'ln_consumption');
+//
+//            // save slope
+//            DB::table('crop_slope')
+//                ->insert([
+//                    'crop_id' => $crop->id,
+//                    'area' => $slopeArea,
+//                    'yield' => $slopeYield,
+//                    'consumption' => $slopeConsumption
+//                ]);
+//
+//            // compute annualize growth rate
+//            $agrArea = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'harvested_area_ha');
+//            $agrYield = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'yield_mt_ha');
+//            $agrConsumption = $this->calculateAnnGrowthRate($crop->id, 'crop_data', 'per_capita_consumption_kg_yr');
+//
+//            // save anualize growth rate
+//            DB::table('crop_annualized_growth_rate')
+//                ->insert([
+//                    'crop_id' => $crop->id,
+//                    'area' => $agrArea,
+//                    'yield' => $agrYield,
+//                    'consumption' => $agrConsumption
+//                ]);
+//
+//        } elseif ($typeCrop->crop_type == 'Non-crops') {
+//
+//            // compute slope
+//            $slopeProduction = $this->calculateSlope($crop->id, 'non_crop_data', 'ln_production');
+//            $slopeConsumption = $this->calculateSlope($crop->id, 'non_crop_data', 'ln_consumption');
+//
+//            // save slope
+//            DB::table('non_crop_slope')
+//                ->insert([
+//                    'crop_id' => $crop->id,
+//                    'production' => $slopeProduction,
+//                    'consumption' => $slopeConsumption
+//                ]);
+//
+//            // compute annualize growth rate
+//            $agrProduction = $this->calculateAnnGrowthRate($crop->id, 'non_crop_data', 'production_mt');
+//            $agrConsumption = $this->calculateAnnGrowthRate($crop->id, 'non_crop_data', 'per_capita_consumption_kg_yr');
+//
+//            // save anualize growth rate
+//            DB::table('non_crop_annualized_growth_rate')
+//                ->insert([
+//                    'crop_id' => $crop->id,
+//                    'production' => $agrProduction,
+//                    'consumption' => $agrConsumption
+//                ]);
+//
+//        }
+////
+////        // import population growth rate
+////        $this->importPopGrowthRate($popGrowthDest, $crop->id, $population);
+////
+////        // create file source for auto-arima model
+//        $this->createAutoArimaSrcData($crop->id, $typeCrop->crop_type);
 
         return "Successfully saved.";
     }
 
     /**
-     * Check commodity data.
-     *
-     * @param [type] $path
-     * @param [type] $cropType
-     */
-    private function checkCommodityData($path, $cropType)
-    {
-        // reading commodity data in csv format
-        $reader = new Csv;
-        $excel = $reader->load(storage_path($path));
-
-        // set active sheet
-        $excel->setActiveSheetIndex(0);
-
-        // sheet dimension
-        $lastColumn = $excel->getActiveSheet()->getHighestColumn();
-        $lastRow = $excel->getActiveSheet()->getHighestRow();
-
-        // columns for "crop" type commodities
-        // -> a. Year
-        // -> b. Area Harvested
-        // -> c. Production
-        // -> d. Per Capita Consumption
-        if ($cropType == 'crops' AND $lastColumn != 'D') {
-            return ['commodity_data' => 'Column out of range.'];
-        } elseif ($cropType == 'crops' AND $lastRow < 6) {
-            return ['commodity_data' => 'Row out of range.'];
-        }
-
-        // columns for "non-crop" type commodities
-        // -> a. Year
-        // -> b. Production
-        // -> c. Per Capita Consumption
-        if ($cropType == 'Non-crops' AND $lastColumn != 'C') {
-            return ['commodity_data' => 'Column out of range.'];
-        } elseif ($cropType == 'Non-crops' AND $lastRow < 6) {
-            return ['commodity_data' => 'Row out of range.'];
-        }
-
-        return false;
-    }
-
-    /**
-     * Check population growth rate.
-     *
-     * @param [type] $path
-     * @param [type] $year
-     */
-    private function checkPopGrowthRate($path, $year)
-    {
-        // reading commodity data in csv format
-        $reader = new Csv;
-        $excel = $reader->load(storage_path($path));
-
-        // set active sheet
-        $excel->setActiveSheetIndex(0);
-
-        // sheet dimension
-        $lastColumn = $excel->getActiveSheet()->getHighestColumn();
-        $lastRow = $excel->getActiveSheet()->getHighestRow();
-
-        // start year with growth rate value must be
-        // year specified in the population plus 1
-        // -> if year is 2015
-        // -> then year 2016 must have growth rate value
-        $y1 = $year + 1;
-        $y2 = $excel->getActiveSheet()->getCell('A2')->getValue();
-        $b2 = $excel->getActiveSheet()->getCell('B2')->getValue();
-
-        // columns and start row for population growth rate
-        if ($lastColumn != 'B') {
-            return ['pop_growth_rate' => 'Column out of range.'];
-        } elseif ($y1 != $y2) {
-            return ['pop_growth_rate' => 'Growth rate must start in year ' . $y1 . '.'];
-        } elseif ($b2 == '' OR $b2 == 0) {
-            return ['pop_growth_rate' => 'Growth rate in first year must be specified.'];
-        }
-
-        return false;
-    }
-
-    /**
-     * Import commodity data.
+     * Import and save commodity data to the database.
      *
      * @param [type] $commodityData
      * @param [type] $cropId
@@ -292,207 +213,80 @@ class InputController extends Controller
      */
     private function importCommodityData($path, $cropId, $rate, $cropType)
     {
-        // reading commodity data in csv format
-        $reader = new Csv;
-        $excel = $reader->load($path);
+        $crop = Crop::find($cropId);
 
-        // set active sheet
-        $excel->setActiveSheetIndex(0);
+        $data = Excel::toArray(new CommodityDataImport, $path);
 
-        // last row of data
-        $lastRow = $excel->getActiveSheet()->getHighestRow();
+        $commodityData = collect([]);
 
-        // loop through all rows starting from row 2
-        for ($row = 2; $row <= $lastRow; $row++) {
+        if ($cropType == 'crops') {
+            foreach ($data[0] as $row) {
+                if (strtolower($row[0]) != 'year') {
+                    $area = (float) str_replace(',','',$row[1]);
+                    $production = (float) str_replace(',','',$row[2]);
+                    $consumption = (float) str_replace(',','',$row[3]);
+                    $yield = round((float) $production / $area, 2) ?? null;
 
-            $year = $excel->getActiveSheet()->getCell('A' . $row)->getValue();
-
-            if ($cropType == 'crops') {
-
-                $area = $excel->getActiveSheet()->getCell('B' . $row)->getValue();
-                $production = $excel->getActiveSheet()->getCell('C' . $row)->getValue();
-                $consumption = $excel->getActiveSheet()->getCell('D' . $row)->getValue();
-
-                // remove commas
-                $area = str_replace(',', '', $area);
-                $production = str_replace(',', '', $production);
-                $consumption = str_replace(',', '', $consumption);
-
-                // if no values, zero is the default
-                $area = ($area == '') ? 0 : $area;
-                $production = ($production == '') ? 0 : $production;
-                $consumption = ($consumption == '') ? 0 : $consumption;
-
-                // compute yield
-                $yield = ($area != 0 AND $production != 0) ? $production / $area : 0;
-
-                // get value applied with conversion rate
-                $areaConversionRate = ($area != 0) ? $area * $rate : 0;
-                $productionConversionRate = ($production != 0) ? $production * $rate : 0;
-                $yieldConversionRate = ($productionConversionRate != 0 AND $areaConversionRate != 0) ? $productionConversionRate / $areaConversionRate : 0;
-
-                // get natural logarithm of a number
-                $lnArea = ($area != 0) ? log($area) : 0;
-                $lnYield = ($production != 0 AND $area != 0) ? log($production / $area) : 0;
-                $lnConsumption = ($consumption != 0) ? log($consumption) : 0;
-
-                // save the data
-                DB::table('crop_data')
-                    ->insert([
-                        'crop_id' => $cropId,
-                        'year' => $year,
-                        'harvested_area_ha' => $area,
-                        'production_mt' => $production,
-                        'yield_mt_ha' => $yield,
-                        'converted_area' => $areaConversionRate,
-                        'converted_production' => $productionConversionRate,
-                        'converted_yield' => $yieldConversionRate,
-                        'per_capita_consumption_kg_yr' => $consumption,
-                        'ln_area' => $lnArea,
-                        'ln_yield' => $lnYield,
-                        'ln_consumption' => $lnConsumption
+                    // apply conversion rate upon validation
+                    $commodityData->push([
+                        'year'          => (int) $row[0],
+                        'harvested_area_ha'          => $area,
+                        'production_mt'    => $production,
+                        'yield_mt_ha'         => $yield,
+                        'converted_area' => $area * $rate,
+                        'converted_production' => $production * $rate,
+                        'converted_yield' => $yield * $rate,
+                        'per_capita_consumption_kg_year'    => $consumption,
+                        'ln_area'       => log($area),
+                        'ln_yield'      => log($yield),
+                        'ln_consumption'=> (float) log($consumption),
                     ]);
+                }
+            }
 
-            } elseif ($cropType == 'Non-crops') {
+            foreach ($commodityData as $item) {
+                $crop->crop_data()->create($item);
+            }
+        } else {
+            // no data
+            foreach ($data[0] as $row) {
+                if (strtolower($row[0]) != 'year') {
+                    $production = (float)  str_replace(',','',$row[1]);
+                    $consumption = (float) str_replace(',','',$row[2]);
 
-                $production = $excel->getActiveSheet()->getCell('B' . $row)->getValue();
-                $consumption = $excel->getActiveSheet()->getCell('C' . $row)->getValue();
-
-                // remove commas
-                $production = str_replace(',', '', $production);
-                $consumption = str_replace(',', '', $consumption);
-
-                // if no values, zero is the default
-                $production = ($production == '') ? 0 : $production;
-                $consumption = ($consumption == '') ? 0 : $consumption;
-
-                // get value applied with conversion rate
-                $productionConversionRate = ($production != 0) ? $production * $rate : 0;
-
-                // get natural logarithm of a number
-                $lnProduction = ($production != 0) ? log($production) : 0;
-                $lnConsumption = ($consumption != 0) ? log($consumption) : 0;
-
-                // save the data
-                DB::table('non_crop_data')
-                    ->insert([
-                        'crop_id' => $cropId,
-                        'year' => $year,
+                    // apply conversion rate upon validation
+                    $commodityData->push([
+                        'year'          => (int) $row[0],
                         'production_mt' => $production,
-                        'converted_production' => $productionConversionRate,
-                        'per_capita_consumption_kg_yr' => $consumption,
-                        'ln_production' => $lnProduction,
-                        'ln_consumption' => $lnConsumption
+                        'converted_production' => $production * $rate,
+                        'per_capita_consumption_kg_year' => $consumption,
+                        'ln_production' => log($production),
+                        'ln_consumption' => log($consumption),
                     ]);
+                }
+            }
 
+            foreach ($commodityData as $item) {
+                $crop->non_crop_data()->create($item);
             }
         }
     }
 
     /**
-     * Calculate slope value.
+     * Call Run LTT backend
      *
-     * @param [type] $cropId
-     * @param [type] $table
-     * @param [type] $column
+     * TODO: ensure that this data is received by backend
      */
-    private function calculateSlope($cropId, $table, $column)
+    private function runModelsInPython($file, $cropId, $conversionRate, $cropType)
     {
-        $sql = "COUNT(*) AS N,
-                SUM(year) AS Sum_X,
-                SUM(year * year) AS Sum_X2,
-                SUM(" . $column . ") AS Sum_Y,
-                SUM(" . $column . " * " . $column . ") AS Sum_Y2,
-                SUM(year * " . $column . ") AS Sum_XY";
+        $url = config('nfcqs.PYTHON_APP');
 
-        $data = DB::table($table)
-            ->selectRaw($sql)
-            ->where('crop_id', $cropId)
-            ->where($column, '<>', 0)
-            ->first();
-
-        if ($data != '' AND $data != null) {
-            return ($data->N * $data->Sum_XY - $data->Sum_X * $data->Sum_Y) / ($data->N * $data->Sum_X2 - $data->Sum_X * $data->Sum_X);
-        } else {
-            return 0;
-        }
-    }
-
-    /**
-     * Calculate annualized growth rate.
-     *
-     * @param [type] $cropId
-     * @param [type] $table
-     * @param [type] $column
-     */
-    private function calculateAnnGrowthRate($cropId, $table, $column)
-    {
-        $count = DB::table($table)
-            ->where('crop_id', $cropId)
-            ->where($column, '<>', 0)
-            ->count();
-
-        $start = DB::table($table)
-            ->where('crop_id', $cropId)
-            ->where($column, '<>', 0)
-            ->orderBy('year', 'ASC')
-            ->first();
-
-        $end = DB::table($table)
-            ->where('crop_id', $cropId)
-            ->where($column, '<>', 0)
-            ->orderBy('year', 'DEC')
-            ->first();
-
-        return (pow(($end->{$column} / $start->{$column}), (1 / $count)) - 1);
-    }
-
-    /**
-     * Import population grwoth rate
-     * and compute population.
-     *
-     * @param [type] $path
-     * @param [type] $cropId
-     * @param [type] $population
-     */
-    private function importPopGrowthRate($path, $cropId, $population)
-    {
-        // reading commodity data in csv format
-        $reader = new Csv;
-        $excel = $reader->load($path);
-
-        // set active sheet
-        $excel->setActiveSheetIndex(0);
-
-        // last row of data
-        $lastRow = $excel->getActiveSheet()->getHighestRow();
-
-        // loop through all rows starting from row 2
-        for ($row = 2; $row <= $lastRow; $row++) {
-
-            $year = $excel->getActiveSheet()->getCell('A' . $row)->getValue();
-            $rate = $excel->getActiveSheet()->getCell('B' . $row)->getValue();
-
-            // compute population
-            $population = $population * (($rate / 100) + 1);
-
-            // insert population growth rate
-            DB::table('population_growth_rate')
-                ->insert([
-                    'crop_id' => $cropId,
-                    'year' => $year,
-                    'rate' => $rate
-                ]);
-
-            // insert computed population
-            DB::table('population')
-                ->insert([
-                    'crop_id' => $cropId,
-                    'year' => $year,
-                    'population' => $population
-                ]);
-
-        }
+        Http::attach('file', $file)->post($url, [
+            'user_id' => auth()->id(),
+            'crop_id' => $cropId,
+            'conversion_rate' => $conversionRate,
+            'crop_type' => $cropType
+        ]);
     }
 
     /**
@@ -503,271 +297,201 @@ class InputController extends Controller
      */
     private function createAutoArimaSrcData($cropId, $cropType)
     {
-        if ($cropType == 'crops') {
+        // this code generates production (actual production, yield, and area only) and consumption data (based on per capita only)
+//        if ($cropType == 'crops') {
 
-            // file path
-            $areaPath = 'uploads/commodity-data/auto-arima/input/area-' . $cropId . '.csv';
-            $yieldPath = 'uploads/commodity-data/auto-arima/input/yield-' . $cropId . '.csv';
-            $consumptionPath = 'uploads/commodity-data/auto-arima/input/consumption-' . $cropId . '.csv';
+//            // file path
+//            $areaPath = 'uploads/commodity-data/auto-arima/input/area-' . $cropId . '.csv';
+//            $yieldPath = 'uploads/commodity-data/auto-arima/input/yield-' . $cropId . '.csv';
+//            $consumptionPath = 'uploads/commodity-data/auto-arima/input/consumption-' . $cropId . '.csv';
+//
+//            // initialize
+//            $spreadsheet = new \PhpOffice\PhpSpreadsheet\Spreadsheet();
+//            $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($spreadsheet);
+//
+//            // create blank files
+//            $writer->save($areaPath);
+//            $writer->save($yieldPath);
+//            $writer->save($consumptionPath);
+//
+//            // get crop data
+//            $data = DB::table('crop_data')
+//                ->where('crop_id', $cropId)
+//                ->orderBy('year', 'ASC')
+//                ->get();
+//
+//            if (sizeof($data) != 0) {
+//
+//                // populate data to area
+//                $reader = new Csv;
+//                $excel = $reader->load($areaPath);
+//                $excel->setActiveSheetIndex(0);
+//                $sheet = $excel->getActiveSheet();
+//                $row = 2;
+//
+//                // set headers
+//                $sheet->setCellValue('A1', 'YEAR');
+//                $sheet->setCellValue('B1', 'AREA');
+//
+//                // set file content
+//                foreach ($data as $key) {
+//
+//                    // only if harvested area is not zero
+//                    if ($key->harvested_area_ha != 0) {
+//
+//                        $sheet->setCellValue('A' . $row, $key->year);
+//                        $sheet->setCellValue('B' . $row, $key->harvested_area_ha);
+//                        $row++;
+//
+//                    }
+//                }
+//
+//                // save the file
+//                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
+//                $writer->setEnclosure('');
+//        	    $writer->save($areaPath);
+//
+//                // populate data to yield
+//                $reader = new Csv;
+//                $excel = $reader->load($yieldPath);
+//                $excel->setActiveSheetIndex(0);
+//                $sheet = $excel->getActiveSheet();
+//                $row = 2;
+//
+//                // set headers
+//                $sheet->setCellValue('A1', 'YEAR');
+//                $sheet->setCellValue('B1', 'YIELD');
+//
+//                // set file content
+//                foreach ($data as $key) {
+//
+//                    // only if yield is not zero
+//                    if ($key->yield_mt_ha != 0) {
+//
+//                        $sheet->setCellValue('A' . $row, $key->year);
+//                        $sheet->setCellValue('B' . $row, $key->yield_mt_ha);
+//                        $row++;
+//
+//                    }
+//                }
+//
+//                // save the file
+//                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
+//                $writer->setEnclosure('');
+//                $writer->save($yieldPath);
 
-            // initialize
-            $spreadsheet = new \PhpOffice\PhpSpreadsheet\Spreadsheet();
-            $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($spreadsheet);
-            
-            // create blank files
-            $writer->save($areaPath);
-            $writer->save($yieldPath);
-            $writer->save($consumptionPath);
-            
-            // get crop data
-            $data = DB::table('crop_data')
-                ->where('crop_id', $cropId)
-                ->orderBy('year', 'ASC')
-                ->get();
-
-            if (sizeof($data) != 0) {
-
-                // populate data to area
-                $reader = new Csv;
-                $excel = $reader->load($areaPath);
-                $excel->setActiveSheetIndex(0);
-                $sheet = $excel->getActiveSheet();
-                $row = 2;
+//                // populate data to consumption
+//                $reader = new Csv;
+//                $excel = $reader->load($consumptionPath);
+//                $excel->setActiveSheetIndex(0);
+//                $sheet = $excel->getActiveSheet();
+//                $row = 2;
 
                 // set headers
-                $sheet->setCellValue('A1', 'YEAR');
-                $sheet->setCellValue('B1', 'AREA');
-
-                // set file content
-                foreach ($data as $key) {
-
-                    // only if harvested area is not zero
-                    if ($key->harvested_area_ha != 0) {
-
-                        $sheet->setCellValue('A' . $row, $key->year);
-                        $sheet->setCellValue('B' . $row, $key->harvested_area_ha);
-                        $row++;
-
-                    }
-                }
-
-                // save the file
-                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
-                $writer->setEnclosure('');
-        	    $writer->save($areaPath);
-
-                // populate data to yield
-                $reader = new Csv;
-                $excel = $reader->load($yieldPath);
-                $excel->setActiveSheetIndex(0);
-                $sheet = $excel->getActiveSheet();
-                $row = 2;
-
-                // set headers
-                $sheet->setCellValue('A1', 'YEAR');
-                $sheet->setCellValue('B1', 'YIELD');
-
-                // set file content
-                foreach ($data as $key) {
-
-                    // only if yield is not zero
-                    if ($key->yield_mt_ha != 0) {
-
-                        $sheet->setCellValue('A' . $row, $key->year);
-                        $sheet->setCellValue('B' . $row, $key->yield_mt_ha);
-                        $row++;
-
-                    }
-                }
-
-                // save the file
-                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
-                $writer->setEnclosure('');
-                $writer->save($yieldPath);
-
-                // populate data to consumption
-                $reader = new Csv;
-                $excel = $reader->load($consumptionPath);
-                $excel->setActiveSheetIndex(0);
-                $sheet = $excel->getActiveSheet();
-                $row = 2;
-
-                // set headers
-                $sheet->setCellValue('A1', 'YEAR');
-                $sheet->setCellValue('B1', 'CONSUMPTION');
-
-                // set file content
-                foreach ($data as $key) {
-
-                    // only if per capita consumption is not zero
-                    if ($key->per_capita_consumption_kg_yr != 0) {
-
-                        $sheet->setCellValue('A' . $row, $key->year);
-                        $sheet->setCellValue('B' . $row, $key->per_capita_consumption_kg_yr);
-                        $row++;
-
-                    }
-                }
-
-                // save the file
-                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
-                $writer->setEnclosure('');
-                $writer->save($consumptionPath);
+//                $sheet->setCellValue('A1', 'YEAR');
+//                $sheet->setCellValue('B1', 'CONSUMPTION');
+//
+//                // set file content
+//                foreach ($data as $key) {
+//
+//                    // only if per capita consumption is not zero
+//                    if ($key->per_capita_consumption_kg_yr != 0) {
+//
+//                        $sheet->setCellValue('A' . $row, $key->year);
+//                        $sheet->setCellValue('B' . $row, $key->per_capita_consumption_kg_yr);
+//                        $row++;
+//
+//                    }
+//                }
+//
+//                // save the file
+//                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
+//                $writer->setEnclosure('');
+//                $writer->save($consumptionPath);
                 
-            }
-        } else {
+//            }
+//        } else {
 
-            // file path
-            $productionPath = 'uploads/commodity-data/auto-arima/input/production-' . $cropId . '.csv';
-            $consumptionPath = 'uploads/commodity-data/auto-arima/input/consumption-' . $cropId . '.csv';
-
-            // initialize
-            $spreadsheet = new \PhpOffice\PhpSpreadsheet\Spreadsheet();
-            $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($spreadsheet);
-            
-            // create blank file
-            $writer->save($productionPath);
-            $writer->save($consumptionPath);
-            
-            // get crop data
-            $data = DB::table('non_crop_data')
-                ->where('crop_id', $cropId)
-                ->orderBy('year', 'ASC')
-                ->get();
-
-            if (sizeof($data) != 0) {
-
-                // populate data to production
-                $reader = new Csv;
-                $excel = $reader->load($productionPath);
-                $excel->setActiveSheetIndex(0);
-                $sheet = $excel->getActiveSheet();
-                $row = 2;
-
-                // set headers
-                $sheet->setCellValue('A1', 'YEAR');
-                $sheet->setCellValue('B1', 'PRODUCTION');
-
-                // set file content
-                foreach ($data as $key) {
-
-                    // only if production is not zero
-                    if ($key->production_mt != 0) {
-
-                        $sheet->setCellValue('A' . $row, $key->year);
-                        $sheet->setCellValue('B' . $row, $key->production_mt);
-                        $row++;
-
-                    }
-                }
-
-                // save the file
-                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
-                $writer->setEnclosure('');
-                $writer->save($productionPath);
-
-                // populate data to consumption
-                $reader = new Csv;
-                $excel = $reader->load($consumptionPath);
-                $excel->setActiveSheetIndex(0);
-                $sheet = $excel->getActiveSheet();
-                $row = 2;
-
-                // set headers
-                $sheet->setCellValue('A1', 'YEAR');
-                $sheet->setCellValue('B1', 'CONSUMPTION');
-
-                // set file content
-                foreach ($data as $key) {
-
-                    // only if per capita consumption is not zero
-                    if ($key->per_capita_consumption_kg_yr != 0) {
-
-                        $sheet->setCellValue('A' . $row, $key->year);
-                        $sheet->setCellValue('B' . $row, $key->per_capita_consumption_kg_yr);
-                        $row++;
-
-                    }
-                }
-
-                // save the file
-                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
-                $writer->setEnclosure('');
-                $writer->save($consumptionPath);
-                
-            }
-        }
-    }
-
-    /**
-     * Upload project report form.
-     *
-     * @param Request $request
-     */
-    public function uploadReportForm(Request $request)
-    {
-        return view('reports.upload');
-    }
-
-    /**
-     * Upload project report.
-     *
-     * @param Request $request
-     */
-    public function uploadReport(Request $request)
-    {
-        $remarks = trim($request->input('remarks'));
-        $file = $request->file('file');
-        
-        // validation
-        $validator = Validator::make(
-            [
-                'remarks' => $remarks,
-                'file' => $file
-            ], [
-                'remarks' => 'required|max:100',
-                'file' => 'required|mimes:doc,dox,xls,xlsx,pdf,ppt,pptx',
-            ]);
-
-        if (!$validator->fails()) {
-
-            // get file extension
-            $path = $_FILES['file']['name'];
-            $ext = pathinfo($path, PATHINFO_EXTENSION);
-
-            // file repo destination
-            $dest = 'uploads/project-reports/' . strtolower(rand() . '.' . $ext);
-            
-            // copy the files
-            copy($file, $dest);
-
-            // insert the file
-            DB::table('project_reports')
-                ->insert([
-                    'filename' => $dest,
-                    'remarks' => $remarks
-                ]);
-
-            return "File successfully uploaded.";
-        } else {
-            return $validator->errors();
-        }
-    }
-
-    /**
-     * List of project reports.
-     *
-     * @param Request $request
-     */
-    public function reportList(Request $request)
-    {
-        $data = DB::table('project_reports')
-            ->orderBy('id', 'DESC')
-            ->get();
-
-        return view('reports.list')->with([
-            'data' => $data
-        ]); 
+//            // file path
+//            $productionPath = 'uploads/commodity-data/auto-arima/input/production-' . $cropId . '.csv';
+//            $consumptionPath = 'uploads/commodity-data/auto-arima/input/consumption-' . $cropId . '.csv';
+//
+//            // initialize
+//            $spreadsheet = new \PhpOffice\PhpSpreadsheet\Spreadsheet();
+//            $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($spreadsheet);
+//
+//            // create blank file
+//            $writer->save($productionPath);
+//            $writer->save($consumptionPath);
+//
+//            // get crop data
+//            $data = DB::table('non_crop_data')
+//                ->where('crop_id', $cropId)
+//                ->orderBy('year', 'ASC')
+//                ->get();
+//
+//            if (sizeof($data) != 0) {
+//
+//                // populate data to production
+//                $reader = new Csv;
+//                $excel = $reader->load($productionPath);
+//                $excel->setActiveSheetIndex(0);
+//                $sheet = $excel->getActiveSheet();
+//                $row = 2;
+//
+//                // set headers
+//                $sheet->setCellValue('A1', 'YEAR');
+//                $sheet->setCellValue('B1', 'PRODUCTION');
+//
+//                // set file content
+//                foreach ($data as $key) {
+//
+//                    // only if production is not zero
+//                    if ($key->production_mt != 0) {
+//
+//                        $sheet->setCellValue('A' . $row, $key->year);
+//                        $sheet->setCellValue('B' . $row, $key->production_mt);
+//                        $row++;
+//
+//                    }
+//                }
+//
+//                // save the file
+//                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
+//                $writer->setEnclosure('');
+//                $writer->save($productionPath);
+//
+//                // populate data to consumption
+//                $reader = new Csv;
+//                $excel = $reader->load($consumptionPath);
+//                $excel->setActiveSheetIndex(0);
+//                $sheet = $excel->getActiveSheet();
+//                $row = 2;
+//
+//                // set headers
+//                $sheet->setCellValue('A1', 'YEAR');
+//                $sheet->setCellValue('B1', 'CONSUMPTION');
+//
+//                // set file content
+//                foreach ($data as $key) {
+//
+//                    // only if per capita consumption is not zero
+//                    if ($key->per_capita_consumption_kg_yr != 0) {
+//
+//                        $sheet->setCellValue('A' . $row, $key->year);
+//                        $sheet->setCellValue('B' . $row, $key->per_capita_consumption_kg_yr);
+//                        $row++;
+//
+//                    }
+//                }
+//
+//                // save the file
+//                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Csv($excel);
+//                $writer->setEnclosure('');
+//                $writer->save($consumptionPath);
+//
+//            }
+//        }
     }
 
     /**
@@ -995,5 +719,58 @@ class InputController extends Controller
             'error' => 0,
             'msg' => "Record successfully deleted."
         ];
+    }
+
+    /**
+     * @param $population
+     * @param $year
+     * @param $file
+     * @return void
+     */
+    private function handlePopulationProjection($population, $cropId, $file): void
+    {
+        // load data from file into array
+        $data = Excel::toArray(new PopulationDataImport, $file);
+
+        // initialize a new array to map year, rate
+        $growthRates = collect([]);
+        foreach ($data[0] as $row) {
+            if (strtolower($row[0]) != 'year') {
+                $growthRates->push([
+                    'year' => $row[0],
+                    'rate' => $row[1]
+                ]);
+            }
+        }
+
+        // initialize population using base population and population projection variables
+        $initialPopulation = $population;
+        $populationProjection = [];
+
+        // iterate over growth rates to project population
+        // set initial population to current population
+        foreach ($growthRates as $item) {
+            $currentPopulation = intval(round($initialPopulation * ($item['rate']/100 + 1)));
+            array_push($populationProjection, [
+                'year'          => $item['year'],
+                'rate'          => $item['rate'],
+                'population'    => $currentPopulation
+            ]);
+            $initialPopulation = $currentPopulation;
+        }
+
+        $crop = Crop::find($cropId);
+
+        foreach ($populationProjection as $item) {
+            $crop->population_growth_rates()->create([
+                'year' => $item['year'],
+                'rate' => $item['rate']
+            ]);
+
+            $crop->populations()->create([
+                'year' => $item['year'],
+                'population' => $item['population']
+            ]);
+        }
     }
 }
