+++ config/nfcqs.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'PYTHON_APP' => env('PYTHON_APP', 'http://localhost:5000'),
+];
