+++ storage/app/public/uploads/commodity-data/ilocos_norte_rice-Y3UcCG.csv
@@ -1,0 +1,33 @@
+﻿Year,Area harvested,Production,Per capita consumption (kg/yr)
+1987,3400.00,5303.00,
+1988,3850.00,5999.00,
+1989,3960.00,6562.00,
+1990,3610.00,5964.00,92.53
+1991,3000.00,5224.00,83.71
+1992,3430.00,7555.00,87.13
+1993,3640.00,9214.00,88.52
+1994,4840.00,9926.00,86.49
+1995,4812.00,8560.00,92.55
+1996,5000.00,9255.00,98.73
+1997,4820.00,9259.00,97.05
+1998,4304.00,7753.00,91.91
+1999,4881.00,9491.00,99.68
+2000,5594.00,11039.00,103.16
+2001,5932.00,11932.00,103.77
+2002,5858.00,13860.00,108.03
+2003,5647.00,12069.00,107.02
+2004,5910.00,13859.00,116.09
+2005,6012.00,14664.00,118.80
+2006,6080.00,15767.00,118.70
+2007,6350.00,16802.00,126.26
+2008,6314.00,17142.00,128.10
+2009,6236.00,16221.00,121.51
+2010,6194.00,15780.00,113.82
+2011,5994.00,15790.00,116.46
+2012,6286.00,16696.00,118.87
+2013,6286.00,17491.00,115.63
+2014,6286.00,17355.00,114.35
+2015,6284.00,18144.00,111.62
+2016,6564.00,18145.00,107.83
+2017,6976.00,20276.00,118.25
+2018,7098.00,19917.68,
