+++ storage/app/public/uploads/pop-growth-rate/ilocos_norte_rice-200ah1.csv
@@ -1,0 +1,26 @@
+﻿YEAR,PGR
+2016,1.77
+2017,1.77
+2018,1.77
+2019,1.77
+2020,1.77
+2021,1.53
+2022,1.53
+2023,1.53
+2024,1.53
+2025,1.53
+2026,1.28
+2027,1.28
+2028,1.28
+2029,1.28
+2030,1.28
+2031,1.07
+2032,1.07
+2033,1.07
+2034,1.07
+2035,1.07
+2036,0.92
+2037,0.92
+2038,0.92
+2039,0.92
+2040,0.92
