--- storage/framework/sessions/DzsdkTtF49AWVznBUdzodCEkczdxXdiK2bopl9kc
@@ -1,1 +1,0 @@
-a:4:{s:6:"_token";s:40:"9pGc5SJag0IUlvySxx4wwh6wlW8iRaJxutSDmOiQ";s:3:"url";a:1:{s:8:"intended";s:17:"http://nfcqs.test";}s:9:"_previous";a:1:{s:3:"url";s:23:"http://nfcqs.test/login";}s:6:"_flash";a:2:{s:3:"old";a:0:{}s:3:"new";a:0:{}}}
