--- storage/framework/sessions/FXYtk6NymDWvs1fzja7O9Ic3HJWmIxseCfBFc2j7
@@ -1,1 +1,0 @@
-a:5:{s:6:"_token";s:40:"Ai8VKt8W34YTnujX53rvmuUoaxERbDL6hE9X3sFW";s:3:"url";a:0:{}s:9:"_previous";a:1:{s:3:"url";s:17:"http://nfcqs.test";}s:6:"_flash";a:2:{s:3:"old";a:0:{}s:3:"new";a:0:{}}s:50:"login_web_59ba36addc2b2f9401580f014c7f58ea4e30989d";i:1;}
