+++ storage/framework/sessions/Wc4yrVALETJXyR2uj9u2tOJKWSnFgxb5tcZtKJDG
@@ -1,0 +1,1 @@
+a:5:{s:6:"_token";s:40:"9QnVJpyQt1hEK6nfmisOcXoHmGacKAsY5fnJvtM9";s:6:"_flash";a:2:{s:3:"old";a:0:{}s:3:"new";a:0:{}}s:3:"url";a:0:{}s:9:"_previous";a:1:{s:3:"url";s:17:"http://nfcqs.test";}s:50:"login_web_59ba36addc2b2f9401580f014c7f58ea4e30989d";i:1;}
